requirements.txt
python-telegram-bot==13.15
telethon==1.28.5
aiohttp==3.8.4
python-dotenv==1.0.0 